aede::commands::play

Trait PlaybackOutput

trait PlaybackOutput: Write {
    // Required methods
    fn pause(&self) -> Result<(), Box<dyn Error>>;
    fn resume(&self) -> Result<(), Box<dyn Error>>;

    // Provided methods
    fn requires_exact(&self) -> bool { ... }
    fn consumed_frames(&self) -> Option<u64> { ... }
    fn write_pcm(
        &mut self,
        _samples: &[f32],
        f32le: &[u8],
        byte_offset: usize,
    ) -> Result<usize> { ... }
    fn write_exact(
        &mut self,
        _samples: &[i32],
        _canonical32: &[u8],
        _byte_offset: usize,
    ) -> Result<usize> { ... }
}

Required Methods

fn pause(&self) -> Result<(), Box<dyn Error>>

fn resume(&self) -> Result<(), Box<dyn Error>>

Provided Methods

fn requires_exact(&self) -> bool

fn consumed_frames(&self) -> Option<u64>

Frames handed to the native device callback, when this output can report them.

fn write_pcm( &mut self, _samples: &[f32], f32le: &[u8], byte_offset: usize, ) -> Result<usize>

Accept PCM from its original samples or encoded transport bytes. The byte offset and return value retain arbitrary partial-write semantics.

fn write_exact( &mut self, _samples: &[i32], _canonical32: &[u8], _byte_offset: usize, ) -> Result<usize>